fix wrong memory size / remove clone
Previously, when trying to launch with mem=2GB (HVM's max heap size), it
took ~3s to run a simple program. The reason was that the mem_size was
in bytes, but HVMC expects the number of 64-bit nodes. So, the default
of 128MB actually allocated 1GB. I also removed an unecessary clone() of
the entire heap that doubled the time.