re-run README example with depth=28
Seems like the single-threaded version was running slower than expected,
possibly due to memory getting full. I've re-ran the test with depth=28,
and the results are more sensible now.

This test was ran on an Apple M3 Max.

The single-core CPU was ran using GCC -O3, with TPC=0 (1 thread).

The multi-core CPU was ran using GCC -O3, with TPC=4 (16 threads).

The multi-core GPU was ran using RTX 4090, with TPB=7 and BPG=7 (128x128 threads).